+++ libraries/rain/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2022-12-06
+- add powerPin to constructor. Optional.
+- add powerPin example
+- add interrupt example (using DO = digital Out)
+- add digital out polling example
+- update readme.md (lower voltage == wetter)
+- fix version number
+- fix percentage() => 0% == DRY and 100% == WET
+- add get- and setDryReference() for better percentage() behaviour.
+- update keywords.txt
+
+
 ## [0.1.1] - 2022-12-03
 - change levels to milliVolts (uint16_t).
 - rewrite setLevel(milliVolts) and getLevel().
@@ -13,7 +25,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add/update examples
 - update keywords.txt
 - add delta() => delta with previous read().
-
 
 ## [0.1.0] - 2022-12-03
 - initial version